backend/requirements.txt
fastapi
uvicorn
motor
google-genai
pydantic
python-dotenv
